pages: unshow Oasis address for secp256k1 accounts
Accounts other than the ed25519 kind aren't usable on the consensus
layer, and other paratimes aren't in widespread use with this
wallet. With this change, we're no longer showing the Oasis address
for non-ed25519 accounts so that people don't mistakenly enter
them into a field that needs a consensus layer address.